Preparing release flows for server and cli
