+++ .github/workflows/deploy.yaml
@@ -20,46 +20,8 @@ concurrency:
   cancel-in-progress: true
 
 jobs:
-  build-console:
-    name: Build console UI
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Checkout & setup
-        uses: actions/checkout@v4
-        with:
-          ref: ${{ inputs.ref }}
-
-      - name: Setup Node.js
-        uses: actions/setup-node@v4
-        with:
-          node-version: 20
-
-      - uses: pnpm/action-setup@v2
-        name: Install pnpm
-        with:
-          version: 8
-          run_install: false
-
-      - name: Install dependencies
-        working-directory: console
-        run: |
-          # install dependencies
-          pnpm install
-
-      - name: Build project
-        working-directory: console
-        run: pnpm run build
-
-      - name: Cache the build
-        uses: actions/cache/save@v3
-        with:
-          path: assets/console
-          key: ${{ inputs.ref }}-console
-
   release-binary:
     name: Build And Release Binaries
-    needs: build-console
     strategy:
       matrix:
         goos: [linux, windows, darwin]
@@ -183,34 +145,26 @@ jobs:
         uses: actions/checkout@v3
         with:
           ref: ${{ inputs.ref }}
-
-      - name: Set up Docker Buildx
-        uses: docker/setup-buildx-action@v3
-
-      - name: Login to Docker Hub registry
-        uses: "docker/login-action@v3"
+      
+      - name: Setup Go
+        uses: actions/setup-go@v4
         with:
-          username: ${{ secrets.DOCKERHUB_USERNAME }}
-          password: ${{ secrets.DOCKERHUB_TOKEN }}
+          go-version: "1.22.x"
+  
+      - uses: ko-build/setup-ko@v0.6
+        env:
+          KO_DOCKER_REPO: docker.io/hiphops/hiphops
 
-      - name: Cache Docker layers
-        uses: actions/cache@v2
-        with:
-          path: /tmp/.buildx-cache
-          key: $${{ runner.os }}-hops
-
-      - name: Build and push
-        uses: docker/build-push-action@v4
-        with:
-          context: .
-          file: deploy/Dockerfile
-          platforms: linux/amd64,linux/arm64
-          push: true
-          tags: hiphops/hiphops:${{ inputs.ref }},hiphops/hiphops:latest
-          cache-from: type=local,src=/tmp/.buildx-cache
-          cache-to: type=local,dest=/tmp/.buildx-cache-new
-
-      - name: Move cache into place
+      - name: Login to Docker Hub Registry
+        env:
+          docker_token: ${{ secrets.DOCKERHUB_TOKEN }}
         run: |
-          rm -rf /tmp/.buildx-cache
-          mv /tmp/.buildx-cache-new /tmp/.buildx-cache
+          echo "${docker_token}" | ko login --username ${{ secrets.DOCKERHUB_USERNAME }} --password-stdin
+          ko build
+      
+      - run: |
+          cd cmd/hiphops
+          ko build -t "hiphops/hiphops:${{ inputs.ref }}" \
+          -t "hiphops/hiphops:latest" \
+          -p "linux/amd64" \
+          -p "linux/arm64"
