+++ deploy/project_template/v1/.gitignore.tmpl
