+++ deploy/project_template/v1/hiphops/nats.conf
