+++ deploy/project_template/v1/pages/css/base.css
@@ -1,0 +1,3 @@
+body {
+  background-color: yellowgreen;
+}
