+++ deploy/project_template/v1/pages/index.md
@@ -1,0 +1,7 @@
+---
+title: Hiphops.io Dev Portal
+description: Automate all the things!
+---
+# Hiphops.io dev portal
+
+Enjoy our automated goodness.
