+++ staticgen/templates/page.html
@@ -8,12 +8,20 @@
   <meta name=description content="{{ .Description }}">
   <link rel=stylesheet href=/css/base.css crossorigin=anonymous>
   
-  <link rel=apple-touch-icon sizes=180x180 href=/apple-touch-icon.png>
-  <link rel=icon type=image/png href=/favicon-32x32.png sizes=32x32>
-  <link rel=icon type=image/png href=/favicon-16x16.png sizes=16x16>
-  <link rel=manifest href=/manifest.json>
-  <link rel=mask-icon href=/safari-pinned-tab.svg color=#0594cb>
-  <meta name=theme-color content="#ffffff">
+  <link
+      rel="apple-touch-icon"
+      sizes="180x180"
+      href="/favicons/apple-touch-icon.png" />
+    <link
+      rel="icon"
+      type="image/png"
+      sizes="32x32"
+      href="/favicons/favicon-32x32.png" />
+    <link
+      rel="icon"
+      type="image/png"
+      sizes="16x16"
+      href="/favicons/favicon-16x16.png" />
   
   <meta property="og:title" content="{{ .Title }}">
   <meta property="og:description" content="{{ .Description }}">
@@ -24,6 +32,23 @@
 </head>
 
 <body>
-{{ .Content -}}
-</body>
+    <div class="sidebar">
+      <ul class="page_group">
+        <ul>
+          <li class="page_title">💎 Page</li>
+
+          <ul class="subpage_list">
+            <li class="subpage_title">Sub-page #1</li>
+            <li class="subpage_title">Sub-page #2</li>
+          </ul>
+        </ul>
+      </ul>
+    </div>
+
+    <div class="markdown_container">
+      <div class="markdown_div">
+        {{ .Content -}}
+      </div>
+    </div>
+  </body>
 </html>
